requirements.txt
typeguard
tqdm
tensorflow-probability
tensorflow-addons
